+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/PreferenceNames.java
@@ -2,7 +2,6 @@ package de.mephisto.vpin.restclient;
 
 public interface PreferenceNames {
   String IGNORED_VALIDATIONS = "ignoredValidations";
-  String IGNORED_MEDIA = "ignoredMedia";
 
   String SYSTEM_NAME = "systemName";
   String SYSTEM_PRESET = "systemPreset";