+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/PreferencesController.java
@@ -1,12 +1,12 @@
 package de.mephisto.vpin.ui;
 
-import de.mephisto.vpin.commons.fx.UIDefaults;
 import de.mephisto.vpin.commons.fx.Features;
+import de.mephisto.vpin.commons.fx.UIDefaults;
 import de.mephisto.vpin.commons.utils.WidgetFactory;
 import de.mephisto.vpin.ui.events.EventManager;
 import de.mephisto.vpin.ui.events.StudioEventListener;
 import de.mephisto.vpin.ui.preferences.PreferenceType;
-import de.mephisto.vpin.ui.preferences.ScreensPreferencesController;
+import de.mephisto.vpin.ui.preferences.UISettingsPreferencesController;
 import de.mephisto.vpin.ui.util.Dialogs;
 import javafx.application.Platform;
 import javafx.event.ActionEvent;
@@ -168,11 +168,6 @@ public class PreferencesController implements Initializable, StudioEventListener
   }
 
   @FXML
-  private void onScreens(ActionEvent event) throws IOException {
-    load("preference-screens.fxml", event);
-  }
-
-  @FXML
   private void onMame(ActionEvent event) throws IOException {
     load("preference-mame.fxml", event);
   }
@@ -297,7 +292,7 @@ public class PreferencesController implements Initializable, StudioEventListener
     lastScreen = screen;
 
     try {
-      FXMLLoader loader = new FXMLLoader(ScreensPreferencesController.class.getResource(screen));
+      FXMLLoader loader = new FXMLLoader(UISettingsPreferencesController.class.getResource(screen));
       Node node = loader.load();
       prefsMain.setCenter(node);
     } catch (Exception e) {